regression-suite-updates: split out shortlist and competition to partial
