+++ tests/NHSD.GPIT.BuyingCatalogue.RegressionTests/Scenarios/ShortlistScenarios.cs
@@ -1,0 +1,54 @@
+﻿using NHSD.GPIT.BuyingCatalogue.RegressionTests.Pages.Admin.ManageSolutions.SolutionApplicationType;
+using NHSD.GPIT.BuyingCatalogue.RegressionTests.Pages.Admin.ManageSolutions.SolutionHostingType;
+using NHSD.GPIT.BuyingCatalogue.RegressionTests.Pages.Shortlist;
+using Xunit;
+
+namespace NHSD.GPIT.BuyingCatalogue.RegressionTests.Scenarios
+{
+    public partial class OrderScenarios
+    {
+        [Fact]
+        [Trait("Shortlist", "Create shortlist")]
+        public void CreateNewShortlistForFoundationCapabilities()
+        {
+            string shortlistName = "Filter by Foundation Capabilities";
+
+            ShortlistPages.CreateNewShortlist();
+
+            ShortlistPages.CreateShortlistForFoundationCapabilities(shortlistName);
+        }
+
+        [Fact]
+        [Trait("Shortlist", "Create shortlist")]
+        public void CreateNewShortlistForFramework()
+        {
+            string shortlistName = "Filter by Framework";
+
+            ShortlistPages.CreateNewShortlist();
+
+            ShortlistPages.CreateShortlistForFramework(shortlistName);
+        }
+
+        [Fact]
+        [Trait("Shortlist", "Create shortlist")]
+        public void CreateNewShortlistForApplicationTypes()
+        {
+            string shortlistName = "Filter by Application type";
+
+            ShortlistPages.CreateNewShortlist();
+
+            ShortlistPages.CreateShortlistForApplicationType(shortlistName, ApplicationTypes.Desktop);
+        }
+
+        [Fact]
+        [Trait("Shortlist", "Create shortlist")]
+        public void CreateNewShortlistForHostingTypes()
+        {
+            string shortlistName = "Filter by Hosting type";
+
+            ShortlistPages.CreateNewShortlist();
+
+            ShortlistPages.CreateShortlistForHostingType(shortlistName, HostingTypes.Public_cloud);
+        }
+    }
+}
